[INTERNAL] sap.m.Table: Remove cell content related line-height styles
 - There were two reasons to define line-height for table cell contents
 1- Specification: Multiline texts must have different line-height. If
this needs to be solved this should be solved in the cell content CSS
and not in the Table.
 2- Pseudo vertical align middle: Normally rows have 3rem height with
Navigation/Selection controls but we had a different vertical alignments
for compact and cozy. So when the vertical alignment is top the normal
texts with default 1rem line-height was too much on the top. But since
we fixed the vertical alignment issues with
Idf831c87349586adb84c3b2d2303d513dc960769 we do not need those content
adapting styles any longer.

If any context related styles needed for table cells then the table
should not unncessarily change cell content styles but those styles must
be added to the cell content style files and tested over there within
the known table cell context class.

Change-Id: I08478adf4a5e3f09ad314f71363a45890ec3b94c