[INTERNAL][FIX]fl + uxap: Correct title on ChangeVisualization
When a section only has a subsection with a title, it takes over the
subsection title. This isn't defined on the ObjectPageSection designtime
metadata, which causes the RTA Change Visualization to use the wrong
title (from the section instead of the subsection).

This fix adds the getLabel() function on the ObjectPageSection
designtime metadata to return the correct title, and also extends the
RTA elementActionTest to include an optional check for the result of
getLabel functions in the designtime metadata.

Change-Id: Ib05a7a71180baa4252abeac255249cd95b906205
BCP: 2370029779