[INTERNAL] sap.m.ObjectMarker: fixed failing qunit test
* There are two internal private controls used by sap.m.ObjectMarker
but these private controls are located in the ObjectMarker.js file.
There is qunit test that tries to load all controls but fails when
tries to load these private controls because they are not separate
modules.

* In the failing test there is a list with such internal controls
which loading have to be skipped, so these two controls are added
in this list.

Change-Id: Ic06a7831778ae309618d3ee404f27c76bfe388ed
BCP: 2370031898