[FIX] MDCTable: Initialization without property infos in odata.v4.TableDelegate
If the property infos are not yet available on initialization, the initialization failed. This regression was introduced in 1.111. Before that, the property infos were always available on initialization. Since 1.111, this is no longer guaranteed due to changes to the property info handling in MDC controls.
Now it works also without property infos, and it correctly updates after the infos are available.

Change-Id: Ie32ff71ae2e7a58b319808f74fc906f2108d1033
BCP: 2370029491