[FIX] DemoKit: Fix failing QUnit for Highlighter
Problem: One of the QUnits for the Highlighter fails on FireFox. The
reason for that is that we are trying to sort object's key in an
ascending order, but those keys are actually numbers as strings. So that
they are sorted incorrectly (for example, 18 comes first, 7 comes
second). It seems Chrome somehow decides to parse those strings as
numbers, when they are retrieved from object's keys, and it orders them
correctly. However, FireFox treats them as strings.

Solution: Parse the object's keys (as we know they are indexes/numbers),
so that we can correctly sort them in ascending order.

BCP: 2370026149
Change-Id: Ieec6541a292012944aea2836349ecb3686afee82
(cherry picked from commit dd333fd55427636ba360b591890d73093ea79809)
CR-Id: 002075125900001232662023