[INTERNAL][FIX]sap.ui.fl: Remove dependency from deleted changes
If condensing triggered the deletion of a change, it still stays in the
initial map of dependencies - but it is no longer relevant for the
dependency checks. This causes an error in cases where controls are
re-created (e.g. SmartLink popover).

This fix ensures that the dependency relevance check simply returns
"false" instead of causing an exception due to the missing change.

Change-Id: I58d75ded4677765e8bd1c279bf18d04b82cd60ef
BCP: 2370029671