[FIX] sap.uxap.ObjectPageSubSection: Fix removeAggregation function
Problem: Due to a previous change, removeAggregation now works with all
possible params - object, string, integer (index). However, it seems, it
was made to work only with objects on purpose (to ignore calls from core
logic and to take into account only self calls).
As a background: previous change was needed, because there is a problem
with flexibility, related to a moment of time, when BlockBases do not
have a parent, as setParent function of BlockBase was overwitten 3 years
ago. It was overrwritten because of logged errors (multiple calls from
core logic alongside calls from ObjectPageSubSection).

Solution: Keep the change in setParent method of BlockBase, so that
there is no moment of time, when the BlockBase does not have parent.
However, restore the previous removeAggregation function of the
ObjectPageSubSection with one addition - checking with what parameter it
is called. It again works only with an object. By checking the
parameter, no error is thrown (so the bug from 3 years ago is also still
prevented).

BCP: 2370036912
Change-Id: I264771c3d1ad2b81519aad1edb576d4ebbe36bff