[INTERNAL] sap.m.* controls: Use public mutators instead of protecteds
 - Two common rendering/invalidation mistakes are fixed.
 - In UI5, the latest moment to adapt a control or its children was the
onBeforeRendering hook of the control. This means control and child
invalidation in the onBeforeRendering hook is allowed and controls
should use public setters instead of protecteds with
bSuppressInvaldiate=true flag.
 - With Ide0d10785c08df032b20dbe5683856a42adfbf9f there is a better
invalidation handling during the rendering of children. This change
allows invalidating child control in a single (parent) rendering cycle
if the rendering of the child control has not been started yet.
 - This change adapts sap.m. controls to use public property setters
when those are used for rendering purposes.
 - With apiVersion=4 optimization controls won't get rendered with their
parent if they are not invalidated therefore the invalidation here is
crucial to be able to convert those control to apiVersion 4

Change-Id: I29e2ee39ca243e1a520f653ed989bfa27275cee4