[INTERNAL][FIX]sap.ui.rta: CViz change state adjustment
The change state "ACTIVATED" in CViz is not accurate and it was leading
to errors - like the wrong state being selected after save. Also in some
spots the states and categories were not being properly referenced.

This change cleans up the states and ensures that they are always
referenced in the code instead of using hard-coded values. It also fixes
the issue with the incorrect state being selected after save.

Change-Id: Ie24ca50f9b21ce6cec9aa313ddc0ef17271525cf
BCP: 2370040603