[INTERNAL][FIX]sap.ui.rta: CViz for changes in templates
Changes done on controls inside templates are not being visualized
correctly and causing an error on the console, as the change handler
information is not being retrieved from the control itself but from
the template host.

This fix ensures that the proper selector is used to retrieve the change
handler data and also displays the indicator on the control that
contains the template.

Change-Id: I9ef945b590179710488995be4bd74cbadccd9448
BCP: 2370045789