[FEATURE] sap.m.Button: implement new accessibleRole property
The property receives values from newly introduced
'ButtonAccessibleRole' enumeration. For now there are only two values
in this enumeration - "Default" and "Link".

When "Default" is set as value of "accessibleRole", the Button appears
without "role" attribute.

When "Link" is set as value of "accessibleRole", the Button appears
with  attribute role="link".

JIRA: BGSOFUIBALKAN-6730

Change-Id: I0426b288e76ba43e374d055f00787454c2a1634e