[FIX] ResourceModel: Fix bundle reload after localizationChange
When the ResourceModel received a bundle instance to be used instead of
creating a new bundle, it was unable to reload the bundle on
localizationChange as it did not have all the required information on
how the instance has been created.
To solve this issue, the ResourceBundle is now able to create a new
instance with the exact same parameters it has been created with.

SNOW: CS20240007487913
Change-Id: I102b7db46786049f15fe6c0aae43af7c3baf82bf
CR-Id: 002075125800003060052024
(cherry picked from commit 9f01b52b05dfa9c4b9efda7770b7f09f3b40b34f)