[FIX] ResourceModel: Fix bundle reload after localizationChange
When the ResourceModel received a bundle instance to be used instead of
creating a new bundle, it was unable to reload the bundle on
localizationChange as it did not have all the required information on
how the instance has been created.
To solve this issue, the ResourceBundle is now able to create a new
instance with the exact same parameters it has been created with.

BCP: 002028376700051991612023
Change-Id: I102b7db46786049f15fe6c0aae43af7c3baf82bf