[INTERNAL] ManagedObject: Invalidation handling
 - By default, ManagedObject#setParent method invalidates the parent but
the bSuppressInvalidate=true flag suppress the invalidation of the
"this" so the child. Since the invalidation happens on the parent the
bSuppressInvalidate flag now suppress the invalidation of the parent.
 - ManagedObject#isInvalidateSuppressed method was not only checking the
own suppress invalidate counter but also checks its ancestors. Such
ancestors check was breaking the locality, for example, the invalidation
of children can be suppressed just because the parent invalidation is
suppressed. This usually happens when a control is inserted as an
aggregation with the bSuppressInvalidate flag. At that point, the
binding propagation starts for the child of the control, and although
property setters of the children are called by the binding, invalidation
does not happen on children since the parent is inserted with
bSuppressInvalidate flag. If the invalidation does not happen on the
children, the apiVersion=4 optimization might skip the rendering of
those controls even though the properties are changed.

Change-Id: I08a84afde7c193e955d3a74eb673bdef9b005455