[INTERNAL] sap.m.IllustratedMessage: Added default aria-labelledby
If illustrationAriaLabelledBy property of the sap.m.IllustratedMessage
is not set, the aria-labelledby attribute on the SVG element should
point to the default title of the control.

Also, the aria-hidden of the SVG element is removed for the purpose of
this task and it will be made optional in the future.

JIRA: BGSOFUIPIRIN-6101
Change-Id: I5461b865e6357f109f703627d4ab7a71e411eba4