[FIX] Accessibility Guide: Flickering in sample prevented
Problem:
Flickering of the DynamicPage due to incorrect styling:
the expected 100% height of the scroll wrapper of
DynamicPage did not apply (i.e. the wrapper did not fit
its parent container, but instead had "auto" height).
As a result, the check for presence of scrollable content
failed.

The 100% height did not apply because the parent container
also had auto height.

Solution:
Set 100% height to the DynamicPage container.

Change-Id: I2d20381aec209cdb839a43d25abf27ef0019c043
BCP: 2370051153