[FIX] sap.f.DynamicPageTitle: Fix moving of actions in RTA mode
Problem: When "actions" button are first renamed, and then moved, the
"move" action is not saved.

Solution: The reason for not saving the "move" action is that ta wrong
source index is provided in the change handle of the DynamicPageTitle.
Is should be the source index (original position), not the target
(new position) index.

BCP: 2380035604
Change-Id: I98a1be80f24e0ab09d569f2c4b8d2ebdde5e3b38