[FIX] sap.m.PlanningCalendar: remove redundant labelled by reference
Issue:
All calendar appointments and rows do have an aria-labelledby reference
to thir corresponding row header instance no matter if the row header
is rendered or now.

Solution:
The aria-labelledby reference is only added when the "showRowHeader"
property is set to "true", which will cause the row headers to get rendered.

BCP: 2280154897
Change-Id: Icaa3ee62b377650fb9fdac32dc424699e9772bcd