[FIX] sap.uxap.ObjectPageSubSection: Min height preserved
Problem:
On very small screens or on high zoom level, the content
of the section with sapUxAPObjectPageSubSectionFitContainer
class may become only partially visible if the height of the
sections container becomes smaller than the content's own
minimal height and the content cannot adapt to the
smaller container height.

Solution
When sizing the subsections, ensure that the height of the
subSection with sapUxAPObjectPageSubSectionFitContainer
class is not smaller than the height of the section's
content.

BCP: 2280192538
Change-Id: I1534f09c45b8a046f0e6b1aee39edad449678918