[FIX] sap.m.Link: adjust click handling
Issue:
The inner click handling was written as
there will always be a DOM reference
corresponding to the link control.
There is a case when "getDomRef" method
would return "undefined" when the link's
DOM is moved into the static area.
The example is with a popover containing
a link and on link click the popover gets
closed and it's corresponding DOM representation
gets moved into the static area.

Solution:
The inner click handler uses the event's target,
in order to find the link's DOM representation
instead of calling the "getDomRef" method.

BCP: 002075129400003578212023
BCP: 002075129500003756512023
BCP: 002075129500003754472023
BCP: 002075129400003632232023
BCP: 002075129500003796592023
BCP: 2370061824
CR-Id: 002075125900001973252023
Change-Id: I2159679e811dfa3b50a758cdc528b8f4632f72cc