[INTERNAL] Demo Kit: runtimeOnly support / removed version handling
The "runtimeOnly support" allows hiding of the Demo Kit content for
concrete patch versions. The VersionNotFound view should properly
display an error message in this case.

The change fixed "removed version" handling bug which display the
version removal message when the version string is found in the current
version, e.g. "1.71.5" is in "1.71.56" and therefore resource not found
in 1.71.56 are displayed as removed. This change introduces a strict
version comparison.

Change-Id: I5f664e76cd6302ea0cf2b8a8ef7c79fd60edf2a7