[INTERNAL] odata.v4.TableDelegate: Set _internalTrigger flag (selectionChange)
With Ia30e080ec5cecc3a3ee7d4e14188d44fedc000f5, the delegate started to clear the selection then updating the binding. This causes issues for the ValueHelp.
To solve these issues, the delegate now sets the flag that is usually set inside the SelectionModelSelection plugin and is forwarded via the MultiSelectionPlugin to selectionChange listeners. This flag indicates to the ValueHelp to basically ignore this event.

BCP: 2370053386
Change-Id: Ib23a78242f0f26883160f705659826b25078f301
CR-Id: 002075125900001997952023