[FIX] sap.uxap.ObjectPageSubSection: fix empty SubSections height
Problem: When there are SubSections, which have only "moreBlocks" and no
"blocks", they have some height (20vh), despite they look empty, before
the "moreBlocks" are expanded.

Solution: The problem comes from a performance optimization, which sets
dummy height of empty SubSections, so that the LazyLoader would not
unstash more SubSections then needed (visible in the view port).
Now, to keep this functionality and to fix the empty SubSections look to
big in case they have only "moreBlocks", but no "blocks", we set this
dummy height (20vh) ONLY to SubSections, which we know that have stashed
content.

BCP: 002075129500003865102023
Change-Id: Ib94b71245a11c3d3b8e50147243b123ab8778c45