[INTERNAL] sap.m.ComboBox: selection reset
Problem:
If invalid key is set as selected, the selected item won't be reset. This is the default behaviour. An extension of the control has a requirement that the selection has to be reset and because of this the whole setSelectedKey setter has to be overwritten.

Solution:
We extract only the logic responsible for the selection reset in a protected predicate method, so it can be easily overwritten by ComboBox extensions.

BCP: 2370054655
Change-Id: I853cb5424bf3358cb343458326b23737d8e8fa34
CR-Id: 002075125800001467542023