[INTERNAL][FIX] sap.ui.fl: enable revert after condense in the same session
- After condensing multiple move changes the revertData is not correct anymore.
During the UIReconstruction simulation the previous index has to be
saved in the revertData of the change using a new function in the condenser information.
This will enable changes to be reverted in the same session as they were condensed
(e.g. LocalReset after save in key user adaptation).

- Additionally more obsolete changes are set to delete with this change:
During simulation of set of reduced changes, some changes might end up
moving a control to the same index. Those changes are now also removed.

- added assertions to the test setup to check for the initial UI after revert

BCP: 2370069612
Change-Id: I7da96bfa979564412b0dded9d7057aec88fa319f