[FIX] sap.uxap.ObjectPageLayout: Subsection sizing corrected
Problem:
The calculation of the size of subsections
with sapUxAPObjectPageSubSectionFitContainer class
is affected by the presence of theme-dependent
margings and paddings.
Failure to take some margin/padding into accunt
resulted in incorrect size.

Solution:
--Refactored the calculation to compute
the spacing around the subSection using
getBoundingClientRect instead of checking
for specific margings and paddings.

BCP: 2280192538
Change-Id: I2a079b65391f7d7202e10616a4bdaab0b10ade1b