[FIX] sap.uxap.ObjectPageLayout: Update scrolling on size change
The page already explicitly disables scrolling when
no scrollable content (to avoid some minor flickering due to
temporary appearance of scrollbar before the layout calculations
have completed).
However, the check for presence of scroll also depends on the size
of the content, so should be actualized on each content resize
Additionally, ensured that the calculation for the height
of the snapped content includes the anchorBar.

BCP: 2280192538

Change-Id: Ic9d4ab5bb7f1e66a777a3d95dc45e43cb99f1a87