[FIX] sap.uxap.ObjectPageSection: Hidden title rename fixed
Problem: User should not be able to rename hidden ObjectPageSections'
titles.

Solution: Previous fix only checks the showTitle property of the
ObjectPageSection to determine whether the title should be enabled for
renaming or not. However, this is not enough for such check, because
ObjectPageLayout has additional logic to show/hide Sections' title,
based not only on the showTitle property, but also on the content of the
Sections (for example, the title of the first Section won't be shown if
there is only one SubSection in it). The correct check is the
_getInternalTitleVisible method, which is also used by the
ObjectPageSection renderer to determine whether the title should be
shown or not.

BCP: 2370069453
Change-Id: I8afdb78db18369a7085fb343b6ae9f552d9e1a17