[FIX] DynamicPageTitle: Double snap content fixed
In DynamicPageTitle, when snappedTitleOnMobile aggregation is set,
it hides the heading, that needs to be snapped, but on initial loading
the header is, in most cases, expanded and that check always passes. But
when an invalidation occurs with snapped header, there should be a check
if the snapped heading needs to be shown, or if it should be hidden,
making space for the snappedTitleOnMobile content.

BCP: 2370055312

Change-Id: I607849e37294c8a7757466c1ca704b81a1f78b06