[INTERNAL] sap.ui.core npm package: Include TestRunner
We are revising the concept of providing the UI5 TestRunner in a
dedicated testRunner middleware as part of UI5 Tooling module
ui5-server. Instead we decided to ship the relevant resources as part of
the sap/ui/core npm package. The middleware may then be deprecated in
the future.

This should resolve issues caused by incompatible (hence UI5
version-dependent) changes in the respective sources.

Change-Id: If9f079d9028fa262a5be53a543defeb9e130e70e