[FIX] sap.uxap.ObjectPageLayout: Fix flickering of table
Problem: Table flickers, when placed in a SubSection with class
sapUxAPObjectPageSubSectionFitContainer.

Solution: The Table has a message strip with top margin, which overlaps
with the bottom margin of the Gird span, and as a result - the BlockBase
seems to slip out of the GridContainer. As it seems to slip out, the
scrollheight of the GridContainer (and all its parents) is higher than
the clientHeight, and the SubSection cannot correct calculates it
height, resulting in infinite loop of setting height to auto/fixed
value.

BCP: 2380049168
Change-Id: I0c26832a1b75d28fceaf4b75862ba8a2f2dce357