[INTERNAL] Table: Introduce sort order "None"
Since the introduction of sap.m.table.columnmenu.Menu, it is possible to
remove sorting of a column. To reflect this change in the API,
sap.ui.table.SortOrder is deprecated and replaced with
sap.ui.core.SortOrder that also has the member "None".

Types in the table have been changed accordingly. The property "sorted"
of Column is deprecated, because the property "sortOrder" can now have
the value sap.ui.core.SortOrder.None.

With Table#sort it is now also possible to unsort a column by passing
sap.ui.core.SortOrder.None.

When a column is unsorted via menu or Table#sort, the "sort" event is
fired with the parameter "sortOrder" set to sap.ui.core.SortOrder.None.

Change-Id: I32655701183b5de5bb598e66236b515f09595f32
Jira: CPOUIFTEAMB-2171