[INTERNAL][FIX] sap.ui.fl: Properly set executeOnSelection flag
Previously, when determining the state of the executeOnSelection flag
of the standard variant, it was checked whether the standard variant
is the current variant. However according to the requirements specified
in TEAM42-9 it is sufficient if the standard variant is the default
variant, regardless of whether it is selected or not.
This commit fixes the flag in certain edge cases, e.g. when the app
state sets the current variant to a different one than the default after
a soft reload.

BCP: 2380052114
Change-Id: Ib9c90dc4d99465b9b6dba52a781be18bf01a1d77