[INTERNAL] DemoKit: Fix version switch dialog data binding
Problem:
The version swtich dialog is not opening and an error is thrown in the console.

The context of the promise resolver was not binded since the merge of the legacy free change (I762b877571e637f548d4fe49c2381d70707c2109)

Solution:
We now bind the context of the promise resolver to the this context of the app controller. Removed data model is also restored.

Change-Id: I2ddee6f95ed7be09e9fa86ab6bdbf05be4897a81
CR-Id: 002075125800001773152023
BCP: 2380058405
(cherry picked from commit fa7071f92718f1e615c5c64e8a0cc59e727d33ce)