[FIX] sap.m.SelectList: Visual documentation alignment
The problem was that the bottom border styling of an
individual list item could not be applied, because the
list item behaves like a table row because of its display mode.

The solution is to apply a collapsing border model to the
container of the list items, so that the bottom border of
the list items can be styled individually.

The issue is present in Quartz and Horizon themes.

Jira: BGSOFUIPIRIN-6112
Change-Id: Ibe36139cfa0e5e90feec97a5df6a3fb51c2d2f7d