[INTERNAL] ui5loader: Start using new ECMAScript features
This change is a "canary bird" to introduce new ES features to the
OpenUI5 / SAPUI5 codebase.

Introduces a Class for the internal Module representation without
changing any other code or logic.

Removes the legacy browser check as outdated browsers such as IE11 are
not able to parse the script anymore so the message wouldn't been shown.

JIRA: CPOUI5FOUNDATION-682
Change-Id: I06856832a9b4a6981a7c48bd2b4f8a6acd0a2c85