[FIX] sap.m.OverflowToolbar: Fix missing Popover menu button
Problem: When OverflowToolbar, which have content with relative width,
is placed in a parent, which at some point receives 0px width and then
another value, the Popover menu button of the OverflowToolbar is not
shown, despite there is not enough space for all its buttons.

Solution: As in this case the parent is invalidated and OFT is also
rerendered (instead of resized), the OFT needs to cache its
controls' sizes again (when there are controls with relative width).
In this case onAfterRendering and _doLayout are called, not _handleResize,
that's why in the _doLayout method now we check whether the control has
been resized in the meanwhile.

BCP: 2370027326
Change-Id: I0cd24b75d1992503545c588cc92ec4c9324f7a88