[INTERNAL] Fix issues with the rerender -> invalidate replacement
When controls are migrated from the deprecated, synchronous rerender()
to the asynchronous invalidate(), some tests fail as they trigger an
invalidation and check the result without waiting for a rendering.

Executing Core.applyChanges() before checking the results fixes such
problems, but should only be done in test code (as it again triggers
a rendering synchronously).

Once ES6+ syntax is enabled, we should provide a test helper like

  await whenNextRenderingFinished();

and roll it out as a replacement for Core.applyChanges();

Change-Id: I10ac37a0f8a46095d144d94093b1a0484becb55a