[INTERNAL][FIX] sap.ui.fl: Fix fake variant creation
In the InitialPrepareFunctions, fake variants are now only created for
the variantManagementReference in other variants or if a variant
dependent control change references the standard variant.
Any other missing variant will be created when the control registers
itself to the model.
Deleted variants will not be added to the map anymore, the changes on
them are ignored.

BCP: 2380058601
Change-Id: I8c29f728f755ff3f03abc763101f4d058f1761d1