[INTERNAL][FIX]sap.ui.fl: Translation of FL Variant Titles
FL Variants Titles can't be assigned to resource models in developer
variants because they are not resolved. This fix resolves the texts when
the VM control is registered to the model.

The change also adjusts the handling of fake variants by making their
handling more generic - since the setTitle change created by the
translation also needs to persist the FlexState invalidation. These new
types of objects are called "runtime-steady".

JIRA: TEAM42-280
Change-Id: I640ea471f3eda15a4aeaa64681887c64f6868ee3