[FIX] Field: Open value help only if editable
by focus or click. The valueHelpRequest to open must only be fired by editable controls, so here it is proved on the control level.

At the end the Delegate-implementation decide if it should really open, but it should only be triggered if the Field is editable.

Change-Id: Ic2c3910f4b4d79dcf88bbac65b7b340274f8bbf4
BCP: 2380108529
CR-Id: 002075125800003791932023
(cherry picked from commit 9e22174b3526fe832dc67068cfa6e4b52175b18c)