[INTERNAL] sap.f.semantic.SemanticPage: Fixed binding issue
Problem: When we have bound content in the SemanticPage, it's not
rendered.

Root cause: Since DynamicPage is a private aggregation of the
SemanticPage, its aggregations are not forwarded automatically.

Solution: In the clone method of the SemanticPage, we set its content
to be the clone of the DynamicPage's content.

BCP: 2280178141
Fixes: https://github.com/SAP/openui5/issues/3652
Change-Id: I5aaccfaa5ddfe19ac08c045484c8af28307397cc