[FIX] sap.uxap.ObjectPageLayout: Lazyloading with pinned header fixed
Problem: When header of ObjectPageLayout is pinned and lazyloading is
enabled, Sections are not loaded.

Solution: The problem comes from the property
_bSupressModifyOnScrollOnce, which is set to "true", when the header
is pinned initially. After the first rendering _onScroll method is
called, but its logic is not executed because of this property. As a
result, all Sections (except the first ones, which are preloaded before
the rendering) are not loaded.
Now, if _onScroll is called from the onAfterRenderingDOMReady method we
execute the whole logic, even if _bSupressModifyOnScrollOnce is "true".

BCP: 2380070339
Change-Id: Iff611c0d49f54e46ced988cd49e67d2106f1623e