[FIX] sap.f.SemanticPage: Correct getter result
Problem:
Failure to iterate the items array returned from
the getters of the "footerCustomActions",
"customShareActions", "titleCustomIconActions"
and "titleCustomIconActions" aggregations,
because the length of the array changes during
iteration.

This is because the getters return a 'live'
array i.e. its length changes if we later
add/remove items to the respective aggregation.

However, the contract of the framework is that
the getter returns an array with a stable
length, i.e. it is not changed if we later
add/remove items to the aggregation.

Solution:
Updated the getters to return a copy of the
internal array for the affected aggregations.

BCP: 2380067826
Change-Id: I875ff9729e8c2fad54536f58e34cd625d724e3c5
Fixes: https://github.com/SAP/openui5/issues/3778