[FEATURE] Field: provide API to check for user interaction
As decission to open value help might depend on user interaction it needs to be known is a field is in the initial state or user interaction has happened.
As the user can change the value and immediately back to the original value no change event might be fired.
As changes could be async the user interation should resolved if the async change had finished.

Change-Id: Ic0fded03302c070dc968b985ced6b1a5b2220cd3
JIRA: CPOUIFDENVER-1909