[INTERNAL][FIX]sap.ui.layout: SimpleForm Change Handlers Binding Info
Some SimpleForm change handlers remove all the elements of the "content"
aggregation, leaving them temporarily without a parent.
The ManagedObject asynchronously checks if the elements still have a
parent and, if not, invalidates their bindings. With the change handlers
now being asynchronous, this is causing unexpected loss of binding info
in some specific cases.

To avoid that, we now put the elements in the "dependents" aggregation
of the SimpleForm, so they always have a parent and don't lose their
binding information.

Change-Id: Ie875b84367e06738bddafdcde15e7bbc4b7b635b
BCP: 002075129400003874402023