[INTERNAL] sap.ui.fl: Update local cache from flex response
Instead of recreating all changes after a save, we use the information
from the storage response to update the existing changes on cache,
so that properties like revertData and applyState are still available.

This change only affects variant-related changes, as UI changes are
still not using a data selector.

Change-Id: I75e05e2c46dcffdecd18703c9579d9116183da1b
JIRA: TEAM42-265