[FEATURE] sap.m.Avatar: New 'enabled' property
There is a new property called 'enabled'. Its default value is 'true'.

When the property is 'true' the Avatar might be interactive and it has a
solid color.

When the property is 'false' the Avatar is non-interactive regardless
of its 'press' handlers. Also its color has opacity.

Additionally:

Added visual changes on interaction. According to the new VD spec there
are styles for hover, active and focus.

JIRA: BGSOFUIPIRIN-6149
Change-Id: If8de6346b3f884f2be8685265af1e62173b4e407