[FIX] P13nFilterPanel: Exclude operations that are not supported are no longer visible
The Problem:
In a special case where MaxLength="1" we are supposed to remove the
"StartsWith", "EndsWith" and "Contains" operators. In order to remove them
we remove all operators and then add all of them without those 3.
This is causing issue because for exclude we are also adding all
operations and some of them are not yet supported for exclude.

The Solution:
Now we remove the previous logic and when we are in the case with
MaxLength="1" we just check and remove the 3 above mentioned operators
before adding them to the dropdown.

BCP: 002075129500004650782023
Change-Id: I5240d03c8bdc33346da67d810950c3b9c66809f6
CR-Id: 002075125800002290262023