[INTERNAL] Shortcut.js: Ensure fieldGroup changes
When a shortcut is triggered, the focus is moved out of the control
to trigger a change (e.g. for Input fields). Afterwards the focus
is set back to the control.
Aas fieldGroup changes are handled async (setTimeout) setting back
the focus must also be async to ensure also a fieldGroup validation.

BCP: 2380057170
CR-Id: 002075125800002676602023
Change-Id: I6848818b44767554bd49f14e8f2a065b0e3d172d