[FIX] sap.m.DatePicker: prevent scrolling on focus
ISSUE: On second opening, the page is scrolled up.

ROOT: The reason this was happening is because by the time we want to
focus element, that is not yet rendered.

SOLUTION: We prevent ItemNavigationScrolling on focus.

BCP: 002075129400001889832023
BCP: 002028376700053150452023
CR-Id: 002075125800002505982023
Change-Id: Ib096b202a74d4bf5ac89d0b60eaab213bad54179
(cherry picked from commit 56ba5fa5b9582dd649272526ae526d8be8058c63)