[INTERNAL] DemoKit: Sample code editing fix
Problem: When switching to another Sample (by previous or next buttons;
by changing the Sample from the left menu, etc.) the ace editor fires
event for document change, which is handled by CodeEditor controller and
FileEditor. As there are events with empty values, the FileEditor
"thinks" the user has deleted the file content, and deletes it (unwanted
behavior). As a result, the file content appears empty, the sample looks
like is it changed (while it is not changed by purpose, it is only
freshly opened and all the content of the first file gets mistakenly
deleted). Also error in Console is thrown (if the first file of the
newly opened sample has different name than the first file of the
previously opened sample, for example, Page.view.xml vs
Button.view.xml).
Also the worst thing that happen: Samples' files are cached every time
you open a Sample, without changing them.

Solution: For now, ignore all empty value events (as it was originally
and was working fine). Only missing functionality - the user will not be
able to delete by himself all the content of a certain file. We can
think of an another solution later. This is not crucial for
functionality.

JIRA: BGSOFUIPIRIN-1422
Change-Id: I943d2dc0c86491bbc3aa88a01dabdf70a6a6d51a