[INTERNAL] sap.tnt.ToolHeader: Fixed UShell styles
this functionality is used only in the Fiori Launchpad Navigation Bar.
It is safe to downport it, as no one else is using it. 

The FLP are using version 1.114 and they want this downport. 

Our team (and the PO) agreed, that we'll downport all the fixes related to the usage of sap.tnt.ToolHeader in FLP to version 1.114.


BCP: 2380071254
Change-Id: I1f1c7bba327a0519593f0bac5874bc6bc81c13da
CR-Id: 002075125800002463452023