[FIX] Support Assistant: Avoid attempts to add the same rules more than once
Whenever a library changes, the RuleSetLoader tries to load the rules of that library and store them in a RuleSet.
Sometimes, this process happens more than once for the same library, due to a race condition.
This results in errors: "Support rule with the id xyz already exists."

Solution:
Load the rules for the same libs only once.

BCP: 2380071560
Change-Id: Ia63b4e3fd3f8195ce00b36eddf0e51e715293d56