[FIX] Field: handle "" and null as "not set" for additionalValue
If a user inputs an invalif value the Condition has no additionalValue part. From outside the model updated the additionalValue to "" (because of the data type). This must not lead to a Fiekd update what would remove an error message.

Simmilar issues could happen if the used model or data type changes  "" to null or undefined or vice versa for valid input.

Change-Id: Ie339223a1ad71bc593f764a4839a86bf3d096ded
BCP: 2380082386