[INTERNAL] sap.ui.fl: Fixed dependency issue during control recreation
If controls get recreated, dependencies also have to be recreated.
To check if the dependencies is still valid, the custom data on the control is used.
Sometimes the control is not available yet, because it might be created by a flex change itself.
This caused an exception. Now a safeguard is added: if the control is not available, the dependency is always still valid.

BCP: 2380085130
CR-Id: 002075125800002625982023
Change-Id: Iaa7f5e31a5d520b82e8b81fca2e1c1f2b8e12350