[INTERNAL] sap.ui.fl: Fixed dependency issue during control recreation
If controls get recreated, dependencies also have to be recreated.
To check if the dependencies is still valid, the custom data on the control is used.
Sometimes the control is not available yet, because it might be created by a flex change itself.
This caused an exception. Now a safeguard is added: if the control is not available, the dependency is always still valid.

BCP: 2380085130
CR-Id: 002075125800002625972023
(cherry picked from commit 36cae9632a8153177d010d1cb554f1e6ff4b97a8)
Change-Id: I15ce9f010a1dd4ca5a02ef4225e1a9b1957214ad