[INTERNAL][FIX] GridTable: Table width is set in AutoRowMode for Safari Browser
In Safari, minWidths do not work for td's. That is why the width property needs to be set on the table elements (column headers and content).
This procedure does not work correctly when the table is in AutoRowMode and has many columns with only minWidth.
The width of the content table is being overwritten due to a second rerendering triggered by the AutoRowMode. (the width is set by the Scrolling extension).
To fix this, the width is set after the explicit content area rendering has been triggered.

BCP: 2380079867
Change-Id: Iaeab9b426f444daf3508470049784a8701c4cf29