[FIX] sap.m.IllustratedMessage: Fixed hard error on invalid property entry
Problem: If we pass undefined to the 'setIllustrationType' public
setter an exception is thrown which breaks the whole application.

Root Cause: In the public setter we call a private function with the
passed value (undefined) which expects a string.

Solution: We call the private function only if we have a string as input.

BCP: 2380086687
Change-Id: Ie21035567a464081c7007c6ec1da2612677c9c10