[INTERNAL] sap.m.IllustratedMessage: Support map based illustration sets
Introduce a way to load assets based on a json mappings, usefull to display different illustrations based on the currently loaded theme.

The new illustrations, meant to be loaded when the horizon theme is selected, are available under the illustrations/v5 directory and the "simple" illustration, meant to be loaded when the horizon HC themes are selected are located in the illustrations/v5/hc directory.

The IllustrationPool now stores the mappings included with the metadata.json of the set.

The Illustration module holds the logic which finds the correct key
in the mappings and builds the symbol id with the new key (if any) as well as the configured prefix. The prefix is used to locate (fetch) the new assets later, as well as to store it in the registry under different key, e.g. v5/tnt-Dot-FaceID.

The IllustrationPool stores the assets ids with their prefix. (v5/, v5/hc/, or empty strings) to ensure that when the theme is switched dynamically, although the assets from the new collections are with the same names, they load as well.

Change-Id: I4dad8a756478e0162963029e26fb6285c05f9535
JIRA: BGSOFUIPIRIN-6143
JIRA: BGSOFUIPIRIN-6180