[INTERNAL] sap.m.Breadcrumbs: prevent too much truncation
Problem: When Breadcrumbs has too long currentLocationText and it is
placed in OverflowToolbar, the last item of the Breadcrumbs is truncated
too much before the control goes in the overflow Popup, which makes it
completely unreadable.

Solution: Change the min-width of the Breadcrumb in OverflowToolbar to
be equal to the width of the currentLocationText item. This way, there
is a little truncation (still keeping the original shrinkable
functionality of the Breadcrumbs), but the control goes in the overflow
Popover earlier.

JIRA: BGSOFUIPIRIN-6073
Change-Id: Id87024020e1fef70f16ce8052045e3a596d3d7be