[FIX] sap.f.DynamicPageTitle: Visibility fixed
Problem: When changing the visibility of sap.f.DynamicPageTitle the
parent control
is not getting invalidated (common rule). Anyway in the case the header
is snapped, the contentwrapper is pushed down by a padding-top calculated
by DynamicPage based on the header and title height, which cannot be
updated, without the control invalidation.

Solution: We add observer to listen to that exact property change in
DynamicPageTitle in order to recalculate dependable metrics.

BCP: 002075129400003724452023

Change-Id: I7fc1dbe0ae09ee4e72d263a116d25aadcef772af